module2/requirements.txt
numpy==1.22.1
paho-mqtt==1.6.1
webcolors==1.11.1